Revert "Make project sd-jwt-vc compliant"
