+++ examples/core-example/kb.ts
@@ -1,12 +1,12 @@
-import { SDJwtVcInstance } from '@sd-jwt/sd-jwt-vc';
+import { SDJwtInstance } from '@sd-jwt/core';
 import { DisclosureFrame } from '@sd-jwt/types';
 import { createSignerVerifier, digest, generateSalt } from './utils';
 
 (async () => {
-  const { signer, verifier } = await createSignerVerifier();
+  const { signer, verifier } = createSignerVerifier();
 
   // Create SDJwt instance for use
-  const sdjwt = new SDJwtVcInstance({
+  const sdjwt = new SDJwtInstance({
     signer,
     signAlg: 'EdDSA',
     verifier,
@@ -33,15 +33,7 @@ import { createSignerVerifier, digest, generateSalt } from './utils';
     custom: 'data',
   };
 
-  const encodedSdjwt = await sdjwt.issue(
-    {
-      iss: 'Issuer',
-      iat: new Date().getTime(),
-      vct: 'https://example.com',
-      ...claims,
-    },
-    disclosureFrame,
-  );
+  const encodedSdjwt = await sdjwt.issue(claims, disclosureFrame);
   console.log('encodedSdjwt:', encodedSdjwt);
   const sdjwttoken = await sdjwt.decode(encodedSdjwt);
   console.log(sdjwttoken);
