+++ examples/core-example/tsconfig.json
