+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@sd-jwt/sd-jwt-vc",
+  "name": "@sd-jwt/sd-jwt",
   "version": "0.3.0",
   "description": "sd-jwt draft 7 implementation in typescript",
   "scripts": {
