+++ packages/node-crypto/src/test/crypto.spec.ts
@@ -1,11 +1,5 @@
 import { describe, expect, test } from 'vitest';
-import { generateSalt, digest, ES256 } from '../index';
-
-// Extract the major version as a number
-const nodeVersionMajor = parseInt(
-  process.version.split('.')[0].substring(1),
-  10,
-);
+import { generateSalt, digest } from '../crypto';
 
 describe('This file is for utility functions', () => {
   test('generateSalt', async () => {
@@ -32,24 +26,5 @@ describe('This file is for utility functions', () => {
     const s1 = await digest(payload, 'SHA512');
     expect(s1).toBeDefined();
     expect(s1.length).toBe(64);
-  });
-
-  (nodeVersionMajor < 20 ? test.skip : test)('ES256', async () => {
-    const { privateKey, publicKey } = await ES256.generateKeyPair();
-    expect(privateKey).toBeDefined();
-    expect(publicKey).toBeDefined();
-    expect(typeof privateKey).toBe('object');
-    expect(typeof publicKey).toBe('object');
-
-    const data =
-      'In cryptography, a salt is random data that is used as an additional input to a one-way function that hashes data, a password or passphrase.';
-    const signer = await ES256.getSigner(privateKey);
-    const signature = await signer(data);
-    expect(signature).toBeDefined();
-    expect(typeof signature).toBe('string');
-
-    const verifier = await ES256.getVerifier(publicKey);
-    const result = await verifier(data, signature);
-    expect(result).toBe(true);
   });
 });
