+++ packages/present/src/test/present.spec.ts
@@ -5,7 +5,7 @@ import {
   presentSync,
   presentableKeys,
   presentableKeysSync,
-} from '../index';
+} from '../present';
 import { decodeSdJwt, decodeSdJwtSync } from '@sd-jwt/decode';
 
 describe('Present tests', () => {
