--- packages/sd-jwt-vc/package.json
@@ -1,65 +1,0 @@
-{
-  "name": "@sd-jwt/sd-jwt-vc",
-  "version": "0.3.0",
-  "description": "sd-jwt draft 7 implementation in typescript",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "exports": {
-    ".": {
-      "import": "./dist/index.mjs",
-      "require": "./dist/index.js"
-    }
-  },
-  "scripts": {
-    "build": "rm -rf **/dist && tsup",
-    "lint": "biome lint ./src",
-    "test": "pnpm run test:node && pnpm run test:browser && pnpm run test:e2e && pnpm run test:cov",
-    "test:node": "vitest run ./src/test/*.spec.ts && vitest run ./src/test/*.spec.ts --environment jsdom",
-    "test:browser": "vitest run ./src/test/*.spec.ts --environment jsdom",
-    "test:e2e": "vitest run ./test/*e2e.spec.ts --environment node",
-    "test:cov": "vitest run --coverage"
-  },
-  "keywords": [
-    "sd-jwt",
-    "sdjwt",
-    "sd-jwt-vc"
-  ],
-  "engines": {
-    "node": ">=16"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/openwallet-foundation-labs/sd-jwt-js"
-  },
-  "author": "Lukas.J.Han <lukas.j.han@gmail.com>",
-  "homepage": "https://github.com/openwallet-foundation-labs/sd-jwt-js/wiki",
-  "bugs": {
-    "url": "https://github.com/openwallet-foundation-labs/sd-jwt-js/issues"
-  },
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@sd-jwt/core": "workspace:*"
-  },
-  "devDependencies": {
-    "@sd-jwt/types": "workspace:*",
-    "@sd-jwt/crypto-nodejs": "workspace:*"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "tsup": {
-    "entry": [
-      "./src/index.ts"
-    ],
-    "sourceMap": true,
-    "splitting": false,
-    "clean": true,
-    "dts": true,
-    "format": [
-      "cjs",
-      "esm"
-    ]
-  },
-  "gitHead": "ded40e4551bde7ae93083181bf26bd1b38bbfcfb"
-}
