--- packages/sd-jwt-vc/test/array_data_types.json
@@ -1,29 +1,0 @@
-{
-  "claims": {
-    "data_types": [null, 42, 3.14, "foo", ["Test"], { "foo": "bar" }]
-  },
-  "disclosureFrame": {
-    "data_types": {
-      "_sd": [0, 1, 2, 3, 4, 5]
-    }
-  },
-  "presentationKeys": [
-    "data_types.0",
-    "data_types.1",
-    "data_types.2",
-    "data_types.3",
-    "data_types.4",
-    "data_types.5"
-  ],
-  "presenatedClaims": {
-    "data_types": [null, 42, 3.14, "foo", ["Test"], { "foo": "bar" }]
-  },
-  "requiredClaimKeys": [
-    "data_types.0",
-    "data_types.1",
-    "data_types.2",
-    "data_types.3",
-    "data_types.4",
-    "data_types.5"
-  ]
-}
