--- packages/sd-jwt-vc/test/array_full_sd.json
@@ -1,21 +1,0 @@
-{
-  "claims": {
-    "is_over": {
-      "13": true,
-      "18": false,
-      "21": false
-    }
-  },
-  "disclosureFrame": {
-    "is_over": {
-      "_sd": ["13", "18", "21"]
-    }
-  },
-  "presentationKeys": ["is_over.18"],
-  "presenatedClaims": {
-    "is_over": {
-      "18": false
-    }
-  },
-  "requiredClaimKeys": ["is_over.18"]
-}
