--- packages/sd-jwt-vc/test/array_in_sd.json
@@ -1,13 +1,0 @@
-{
-  "claims": {
-    "sd_array": ["32", "23"]
-  },
-  "disclosureFrame": {
-    "_sd": ["sd_array"]
-  },
-  "presentationKeys": ["sd_array"],
-  "presenatedClaims": {
-    "sd_array": ["32", "23"]
-  },
-  "requiredClaimKeys": ["sd_array"]
-}
