--- packages/sd-jwt-vc/test/array_nested_in_plain.json
@@ -1,20 +1,0 @@
-{
-  "claims": {
-    "nested_array": [["foo", "bar"], ["baz", "qux"]]
-  },
-  "disclosureFrame": {
-    "nested_array": {
-      "0": {
-        "_sd": [0, 1]
-      },
-      "1": {
-        "_sd": [0, 1]
-      }
-    }
-  },
-  "presentationKeys": ["nested_array.0.0", "nested_array.1.1"],
-  "presenatedClaims": {
-    "nested_array": [["foo"], ["qux"]]
-  },
-  "requiredClaimKeys": ["nested_array.0.0", "nested_array.1.0"]
-}
