--- packages/sd-jwt-vc/test/array_none_disclosed.json
@@ -1,17 +1,0 @@
-{
-  "claims": {
-    "is_over": {
-      "13": false,
-      "18": true,
-      "21": false
-    }
-  },
-  "disclosureFrame": {
-    "is_over": {
-      "_sd": ["13", "18", "21"]
-    }
-  },
-  "presentationKeys": [],
-  "presenatedClaims": {},
-  "requiredClaimKeys": []
-}
