--- packages/sd-jwt-vc/test/array_of_nulls.json
@@ -1,15 +1,0 @@
-{
-  "claims": {
-    "null_values": [null, null, null, null]
-  },
-  "disclosureFrame": {
-    "null_values": {
-      "_sd": [1, 2]
-    }
-  },
-  "presentationKeys": [],
-  "presenatedClaims": {
-    "null_values": [null, null]
-  },
-  "requiredClaimKeys": ["null_values.0", "null_values.1"]
-}
