--- packages/sd-jwt-vc/test/array_of_objects.json
@@ -1,58 +1,0 @@
-{
-  "claims": {
-    "addresses": [
-      {
-        "street": "123 Main St",
-        "city": "Anytown",
-        "state": "NY",
-        "zip": "12345",
-        "type": "main_address"
-      },
-      {
-        "street": "456 Main St",
-        "city": "Anytown",
-        "state": "NY",
-        "zip": "12345",
-        "type": "secondary_address"
-      }
-    ],
-    "array_with_one_sd_object": {
-      "foo": "bar"
-    }
-  },
-  "disclosureFrame": {
-    "addresses": {
-      "_sd": [1]
-    },
-    "array_with_one_sd_object": {
-      "_sd": ["foo"]
-    }
-  },
-  "presentationKeys": ["addresses.1", "array_with_one_sd_object.foo"],
-  "presenatedClaims": {
-    "addresses": [
-      {
-        "street": "123 Main St",
-        "city": "Anytown",
-        "state": "NY",
-        "zip": "12345",
-        "type": "main_address"
-      },
-      {
-        "street": "456 Main St",
-        "city": "Anytown",
-        "state": "NY",
-        "zip": "12345",
-        "type": "secondary_address"
-      }
-    ],
-    "array_with_one_sd_object": {
-      "foo": "bar"
-    }
-  },
-  "requiredClaimKeys": [
-    "addresses.0.type",
-    "addresses.1.city",
-    "array_with_one_sd_object.foo"
-  ]
-}
