--- packages/sd-jwt-vc/test/array_of_scalars.json
@@ -1,15 +1,0 @@
-{
-  "claims": {
-    "nationalities": ["US", "CA", "DE"]
-  },
-  "disclosureFrame": {
-    "nationalities": {
-      "_sd": [0, 1]
-    }
-  },
-  "presentationKeys": ["nationalities.1"],
-  "presenatedClaims": {
-    "nationalities": ["CA", "DE"]
-  },
-  "requiredClaimKeys": ["nationalities.0", "nationalities.1"]
-}
