--- packages/sd-jwt-vc/test/array_recursive_sd.json
@@ -1,35 +1,0 @@
-{
-  "claims": {
-    "array_with_recursive_sd": [
-      "boring",
-      {
-        "foo": "bar",
-        "baz": {
-          "qux": "quxx"
-        }
-      },
-      ["foo", "bar"]
-    ],
-    "test2": ["foo", "bar"]
-  },
-  "disclosureFrame": {
-    "array_with_recursive_sd": {
-      "_sd": [1],
-      "1": {
-        "_sd": ["baz"]
-      },
-      "2": {
-        "_sd": [0, 1]
-      }
-    },
-    "test2": {
-      "_sd": [0, 1]
-    }
-  },
-  "presentationKeys": [],
-  "presenatedClaims": {
-    "array_with_recursive_sd": ["boring", []],
-    "test2": []
-  },
-  "requiredClaimKeys": ["array_with_recursive_sd.0", "test2"]
-}
