--- packages/sd-jwt-vc/test/array_recursive_sd_some_disclosed.json
@@ -1,55 +1,0 @@
-{
-  "claims": {
-    "array_with_recursive_sd": [
-      "boring",
-      {
-        "foo": "bar",
-        "baz": {
-          "qux": "quxx"
-        }
-      },
-      ["foo", "bar"]
-    ],
-    "test2": ["foo", "bar"]
-  },
-  "disclosureFrame": {
-    "array_with_recursive_sd": {
-      "1": {
-        "_sd": ["baz"]
-      },
-      "2": {
-        "_sd": [0, 1]
-      }
-    },
-    "test2": {
-      "_sd": [0, 1]
-    }
-  },
-  "presentationKeys": [
-    "array_with_recursive_sd.1.baz",
-    "array_with_recursive_sd.2.1",
-    "test2.0",
-    "test2.1"
-  ],
-  "presenatedClaims": {
-    "array_with_recursive_sd": [
-      "boring",
-      {
-        "foo": "bar",
-        "baz": {
-          "qux": "quxx"
-        }
-      },
-      ["bar"]
-    ],
-    "test2": ["foo", "bar"]
-  },
-  "requiredClaimKeys": [
-    "array_with_recursive_sd.1",
-    "array_with_recursive_sd.2",
-    "array_with_recursive_sd.1.baz",
-    "array_with_recursive_sd.2.1",
-    "test2.0",
-    "test2.1"
-  ]
-}
