--- packages/sd-jwt-vc/test/complex.json
@@ -1,43 +1,0 @@
-{
-  "claims": {
-    "firstname": "John",
-    "lastname": "Doe",
-    "ssn": "123-45-6789",
-    "id": "1234",
-    "data": {
-      "firstname": "John",
-      "lastname": "Doe",
-      "ssn": "123-45-6789",
-      "list": [{ "r": "1" }, "b", "c"]
-    },
-    "data2": {
-      "hi": "bye"
-    }
-  },
-  "disclosureFrame": {
-    "_sd": ["firstname", "id", "data2"],
-    "data": {
-      "_sd": ["list"],
-      "_sd_decoy": 2,
-      "list": {
-        "_sd": [0, 2],
-        "_sd_decoy": 1,
-        "0": {
-          "_sd": ["r"]
-        }
-      }
-    },
-    "data2": {
-      "_sd": ["hi"]
-    }
-  },
-  "presentationKeys": ["firstname", "id"],
-  "presenatedClaims": {
-    "lastname": "Doe",
-    "ssn": "123-45-6789",
-    "data": { "firstname": "John", "lastname": "Doe", "ssn": "123-45-6789" },
-    "id": "1234",
-    "firstname": "John"
-  },
-  "requiredClaimKeys": ["firstname", "id", "data.ssn"]
-}
