--- packages/sd-jwt-vc/test/json_serialization.json
@@ -1,44 +1,0 @@
-{
-  "claims": {
-    "sub": "john_deo_42",
-    "given_name": "John",
-    "family_name": "Deo",
-    "email": "johndeo@example.com",
-    "phone": "+1-202-555-0101",
-    "address": {
-      "street_address": "123 Main St",
-      "locality": "Anytown",
-      "region": "Anystate",
-      "country": "US"
-    },
-    "birthdate": "1940-01-01"
-  },
-  "disclosureFrame": {
-    "_sd": [
-      "sub",
-      "given_name",
-      "family_name",
-      "email",
-      "phone",
-      "address",
-      "birthdate"
-    ]
-  },
-  "presentationKeys": ["given_name", "family_name", "address"],
-  "presenatedClaims": {
-    "given_name": "John",
-    "family_name": "Deo",
-    "address": {
-      "street_address": "123 Main St",
-      "locality": "Anytown",
-      "region": "Anystate",
-      "country": "US"
-    }
-  },
-  "requiredClaimKeys": [
-    "given_name",
-    "family_name",
-    "address.street_address",
-    "address.country"
-  ]
-}
