--- packages/sd-jwt-vc/test/no_sd.json
@@ -1,36 +1,0 @@
-{
-  "claims": {
-    "recursive": [
-      "boring",
-      {
-        "foo": "bar",
-        "baz": {
-          "qux": "quxx"
-        }
-      },
-      ["foo", "bar"]
-    ],
-    "test2": ["foo", "bar"]
-  },
-  "disclosureFrame": {},
-  "presentationKeys": [],
-  "presenatedClaims": {
-    "recursive": [
-      "boring",
-      {
-        "foo": "bar",
-        "baz": {
-          "qux": "quxx"
-        }
-      },
-      ["foo", "bar"]
-    ],
-    "test2": ["foo", "bar"]
-  },
-  "requiredClaimKeys": [
-    "recursive.0",
-    "recursive.1.baz.qux",
-    "recursive.2.1",
-    "test2.1"
-  ]
-}
