--- packages/sd-jwt-vc/vitest.config.mts
@@ -1,4 +1,0 @@
-// vite.config.ts
-import { allEnvs } from '../../vitest.shared';
-
-export default allEnvs;
