+++ packages/types/src/type.ts
@@ -1,6 +1,7 @@
 export const SD_SEPARATOR = '~';
 export const SD_LIST_KEY = '...';
 export const SD_DIGEST = '_sd';
+export const SD_JWT_TYP = 'sd-jwt';
 export const SD_DECOY = '_sd_decoy';
 export const KB_JWT_TYP = 'kb+jwt';
